+++ src/main/java/org/opensearch/neuralsearch/constants/SemanticFieldConstants.java
@@ -53,4 +53,10 @@ public class SemanticFieldConstants {
      * knn_vector field.
      */
     public static final String DENSE_EMBEDDING_CONFIG = "dense_embedding_config";
+
+    /**
+     * Name of the field for sparse encoding config. The config will be used to control how to do sparse encoding.
+     * {@link org.opensearch.neuralsearch.mapper.dto.SparseEncodingConfig}
+     */
+    public static final String SPARSE_ENCODING_CONFIG = "sparse_encoding_config";
 }